# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pa Emmanuel Ayodele Abatan memorial invitation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /event:
    get:
      operationId: getEvent
      tags: [event]
      summary: Get public memorial event
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [admin]
      summary: Update memorial event settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programme:
    get:
      operationId: listProgramme
      tags: [event]
      summary: List funeral programme items
      responses:
        "200":
          description: Programme items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgrammeItem"
    post:
      operationId: createProgrammeItem
      tags: [admin]
      summary: Add programme item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgrammeItemInput"
      responses:
        "201":
          description: Created programme item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgrammeItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programme/{id}:
    patch:
      operationId: updateProgrammeItem
      tags: [admin]
      summary: Update programme item
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgrammeItemUpdate"
      responses:
        "200":
          description: Updated programme item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgrammeItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteProgrammeItem
      tags: [admin]
      summary: Delete programme item
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /public/invitations/{token}:
    get:
      operationId: getPublicInvitation
      tags: [invitations]
      summary: Get a guest invitation by secure token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicInvitation"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/invitations/{token}/rsvp:
    post:
      operationId: submitRsvp
      tags: [invitations]
      summary: Submit RSVP for a guest invitation
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpInput"
      responses:
        "200":
          description: Updated invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicInvitation"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get dashboard counts and recent activity
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/guests:
    get:
      operationId: listGuests
      tags: [admin]
      summary: List and filter guests
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, confirmed, checked_in, disabled]
        - name: rsvp
          in: query
          schema:
            type: string
            enum: [all, yes, no, pending]
      responses:
        "200":
          description: Guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminGuest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createGuest
      tags: [admin]
      summary: Create a guest and invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Guest and invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGuest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/guests/import:
    post:
      operationId: importGuests
      tags: [admin]
      summary: Import guests from CSV text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestImportInput"
      responses:
        "200":
          description: Import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestImportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/guests/{id}:
    get:
      operationId: getGuest
      tags: [admin]
      summary: Get one guest
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGuest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGuest
      tags: [admin]
      summary: Update guest and invitation
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGuest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteGuest
      tags: [admin]
      summary: Delete guest and invitation
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/guests/{id}/disable:
    post:
      operationId: disableInvitation
      tags: [admin]
      summary: Disable an invitation
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGuest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/guests/{id}/enable:
    post:
      operationId: enableInvitation
      tags: [admin]
      summary: Enable an invitation
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGuest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/guests/{id}/regenerate:
    post:
      operationId: regenerateInvitation
      tags: [admin]
      summary: Generate a new invitation code and token
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGuest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/invitations/lookup:
    post:
      operationId: lookupInvitation
      tags: [admin]
      summary: Validate an invitation code before admission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationLookupInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationLookupResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/invitations/{id}/admit:
    post:
      operationId: admitInvitation
      tags: [admin]
      summary: Atomically admit an invitation once
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Admission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /admin/audit:
    get:
      operationId: listAuditLog
      tags: [admin]
      summary: List recent admin actions
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, deceasedName, eventTitle, waykeepDate, burialDate, venue, dressCode]
      properties:
        id: { type: number }
        deceasedName: { type: string }
        eventTitle: { type: string }
        waykeepDate: { type: string }
        burialDate: { type: string }
        year: { type: ["number", "null"] }
        venue: { type: string }
        dressCode: { type: string }
        biography: { type: ["string", "null"] }
        tribute: { type: ["string", "null"] }
        importantInformation: { type: ["string", "null"] }
        directions: { type: ["string", "null"] }
        contactInformation: { type: ["string", "null"] }
        mapUrl: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        backgroundImageUrl: { type: ["string", "null"] }
...
    ProgrammeItem:
      type: object
      required: [id, title, date, sortOrder]
      properties:
        id: { type: number }
        title: { type: string }
        date: { type: string }
        time: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        sortOrder: { type: number }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...